refactor(server): #125 — move embedded GUI handler into ws.rs
